+++ README.md
@@ -76,15 +76,15 @@ shows most of the commands.:
 
 **Instead of Ctrl-letter (e.g. Ctrl-Q), Alt-letter (e.g. Alt-Q) or ESC-letter (e.g. ESC followed by Q) can be used, avoiding conflicts with key binding of some terminal emulators.**
 
-The editor is contained in the files pye_gen.py and pye.py. Start pye from the REPL prompt
+The editor is contained in the files pye_gen.py and pye_core.py. Start pye from the REPL prompt
 e.g. with 
 
-    from pye_gen import pye 
+    from pye import pye 
     res = pye(object_1, object_2, ..[, tabsize=n][, undo=n]) 
 
-You may also use the combined version pye_mp.py resp. pye_mp.mpy:
+You may also use the combined version pye.py resp. pye.mpy:
 
-    from pye_mp import pye  
+    from pye import pye  
     res = pye(object_1, object_2, ..[, tabsize=n][, undo=n]) 
 
 If object_n is a string, it's considered as the name of a file to be edited
@@ -129,7 +129,7 @@ when called with CPython also accepts data from a pipe or redirection.
 ## Files
 
 ### Base files
-- pye.py: Core file with comments, intended to be the same for all platforms 
+- pye_core.py: Core file with comments, intended to be the same for all platforms 
 - pye_ux.py: Front-end for Linux CPython and Linux MicroPython
 - pye_gen.py: Front-end for Micropython and Circuitpython modules with I/O through
 sys.stdout and sys.stdin.
@@ -143,7 +143,7 @@ the front-end files, using cat and sed.
 - README.md: This one
 
 ### Derived files
-- pye_mp.py, pye_mp.mpy: Condensed source files of pye.py + pye_gen.py for
+- pye.py, pye.mpy: Condensed source files of pye_core.py + pye_gen.py for
 all MicroPython boards. In order to use it on an board with small memory
 like the esp8266, you have to put pye_mp.py into the directory esp8266/modules,
 esp32/modules or stm32/modules (micropython.org) or esp32/frozen (pycom.io) and
